Original change: https://android-review.googlesource.com/c/platform/external/perfetto/+/2982252

+++ ui/release/build_all_channels.py
@@ -99,6 +99,8 @@ def main():
   parser = argparse.ArgumentParser()
   parser.add_argument('--upload', action='store_true')
   parser.add_argument('--tmp', default='/tmp/perfetto_ui')
+  parser.add_argument('--branch_only')
+
   args = parser.parse_args()
 
   # Read the releases.json, which maps channel names to git refs, e.g.:
@@ -107,20 +109,24 @@ def main():
   with open(pjoin(CUR_DIR, 'channels.json')) as f:
     channels = json.load(f)['channels']
 
+  if args.branch_only:
+    channels = [{'name': 'branch', 'rev': args.branch_only}]
+
   merged_dist_dir = pjoin(args.tmp, 'dist')
   check_call_and_log(['rm', '-rf', merged_dist_dir])
   shutil.os.makedirs(merged_dist_dir)
   channel_map = build_all_channels(channels, args.tmp, merged_dist_dir)
 
-  print('Updating index in ' + merged_dist_dir)
-  with open(pjoin(merged_dist_dir, 'index.html'), 'r+') as f:
-    index_html = f.read()
-    f.seek(0, 0)
-    f.truncate()
-    index_html = re.sub(r"data-perfetto_version='[^']*'",
-                        "data-perfetto_version='%s'" % json.dumps(channel_map),
-                        index_html)
-    f.write(index_html)
+  if not args.branch_only:
+    print('Updating index in ' + merged_dist_dir)
+    with open(pjoin(merged_dist_dir, 'index.html'), 'r+') as f:
+      index_html = f.read()
+      f.seek(0, 0)
+      f.truncate()
+      index_html = re.sub(
+          r"data-perfetto_version='[^']*'",
+          "data-perfetto_version='%s'" % json.dumps(channel_map), index_html)
+      f.write(index_html)
 
   if not args.upload:
     return
